requirements.txt
einops
# gensim
h5py
matplotlib
numpy
omegaconf
Pillow
pytorch_lightning
tokenizers
torch
torchmetrics
tqdm
wandb
tensorboard
torchtext
torchvision
scikit-learn
ipython
pandas
git+https://github.com/dr-aheydari/SoftAdapt.git